+++ packages/lib/src/propTransform/index.ts
@@ -1,2 +1,4 @@
-export * from "./asDate"
-export * from "./propTransform"
+export { PropTransform, propTransform, PropTransformDecorator } from "./propTransform"
+export * from "./transformArrayAsSet"
+export * from "./transformStringAsDate"
+export * from "./transformTimestampAsDate"