+++ packages/lib/src/snapshot/SnapshotOf.ts
@@ -1,6 +1,6 @@
 import { Frozen, frozenKey } from "../frozen/Frozen"
 import { modelIdKey, modelTypeKey } from "../model"
-import { AnyModel, ModelCreationData, ModelData } from "../model/BaseModel"
+import { AnyModel, ModelPropsCreationData, ModelPropsData } from "../model/BaseModel"
 import { ArraySet, ObjectMap } from "../wrappers"
 
 // snapshot out
@@ -14,7 +14,7 @@ export type SnapshotOutOfObject<T extends { [k: string]: any }> = {
   [k in keyof T]: SnapshotOutOf<T[k]> extends infer R ? R : never
 }
 
-export type SnapshotOutOfModel<M extends AnyModel> = SnapshotOutOfObject<ModelData<M>> & {
+export type SnapshotOutOfModel<M extends AnyModel> = SnapshotOutOfObject<ModelPropsData<M>> & {
   [modelTypeKey]: string
   [modelIdKey]: string
 }
@@ -76,7 +76,7 @@ export type SnapshotInOfObject<T extends { [k: string]: any }> = {
 }
 
 export type SnapshotInOfModel<M extends AnyModel> = SnapshotInOfObject<
-  M extends { fromSnapshot(sn: infer S): any } ? S : ModelCreationData<M>
+  M extends { fromSnapshot(sn: infer S): any } ? S : ModelPropsCreationData<M>
 > & {
   [modelTypeKey]: string
   [modelIdKey]: string